LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-conditionals-expected.txt


PASS Basic @supports
PASS Basic @supports (false)
PASS Nested @supports
PASS Nested @supports (false)
PASS Inconsequential conditional
FAIL @supports with locals assert_equals: expected "1 20 3" but got "var(--x) var(--y) var(--z)"
PASS Basic @media
PASS Basic @media (false)
PASS Nested @media
PASS Nested @media (false)
FAIL Locals within @media assert_equals: expected "PASS" but got "var(--x)"
FAIL @supports within @media assert_equals: expected "PASS" but got "var(--x)"
FAIL @media within @supports assert_equals: expected "PASS" but got "var(--x)"
PASS Basic @container
FAIL Basic @container (false) assert_equals: expected "PASS" but got "FAIL"
FAIL Nested @container assert_equals: expected "PASS" but got "FAIL"
FAIL Nested @container (false) assert_equals: expected "PASS" but got "FAIL"
FAIL Locals within @container assert_equals: expected "PASS" but got "var(--x)"
FAIL @supports within @container assert_equals: expected "PASS" but got "var(--x)"
FAIL @container within @supports assert_equals: expected "PASS" but got "var(--x)"
FAIL @container, @media, @supports assert_equals: expected "PASS" but got "var(--x)"
FAIL @supports, @media, @container assert_equals: expected "PASS" but got "var(--x)"
